+++ back/src/bsvhu/database.ts
@@ -1,7 +1,7 @@
 import { FormNotFound } from "../forms/errors";
 import prisma from "../prisma";
 
-export async function getFormOrFormNotFound(id: string) {
+export async function getBsvhuOrNotFound(id: string) {
   const form = await prisma.bsvhu.findUnique({
     where: { id }
   });