+++ back/src/common/constants/index.ts
@@ -1,3 +1,4 @@
 export * from "./PROCESSING_OPERATIONS";
 export * from "./WASTES";
 export * from "./DASRI_CONSTANTS";
+export * from "./ROAD_CONTROL";